KVM: ARM: selftest: Add stub MMIO register to force a guest exit
For some tests we would like to force a guest exit to stress the
context switching code. This adds a MMIO register that when accessed
ignores any read or write, and implements a vm_exit() function for
guests to call when they need to force an exit to the host.