KVM: ARM: Factor out a15 coprocessor emulation
Factor out A15 coprocessor emulation into separate file.  This should
make it much cleaner to add new cores later on and makes the code more
readable as you can clearly tell which parts can be found in the A15 TRM
and which parts are in the ARM arm.

Warning: This breaks compatibility with QEMU as the VCPU init now takes
a different value for initializing the CPU.  Updated QEMU branch can be
found here:
    git@github.com:virtualopensystems/qemu.git kvm-arm

Cc: Peter Maydell <peter.maydell@linaro.org>
Cc: Will Deacon <will.deacon@arm.com>
Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>