kvm tools: Drop lchown() calls from 9p
Creating new files in a guest:

  echo "hello, world" > hello.txt

results in EPERM on my machine.

The problem is that virtio_p9_create() calls lchown() on UID zero (we
are logged in as root in the guest) which obviously doesn't work. In
fact, ownership changes can't work unless we have some sort of uid/gid
mapping. Therefore, drop the calls to lchown().

Signed-off-by: Pekka Enberg <penberg@kernel.org>