ARM: KVM: introduce {save, restore}_host_regs macros
Host and guest registers are saved/restored using the same set
of macros. This makes it quite difficult to modify the layout of
struct kvm_cpu_regs without impacting the stack layout as well
(which is where the host registers are saved).

Introduce a separate set of macros that are used to save and
restore the host registers using the stack, allowing the guest
macros to be rewritten independently in a subsequent patch.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Christoffer Dall <cdall@cs.columbia.edu>