ARM: KVM: unify HIFAR and HDFAR
There is strictly no reason to have two fields to describe the
faulting VA - we only get one fault at a time, and the 64bit
architecture unifies it for that very reason. We have to rely
on HSR to find out how to interpret the data anyway.

Squash the two fields into one, and perform the associated cleanup.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Christoffer Dall <cdall@cs.columbia.edu>