KVM: ARM: Rework ARM-mode mmio instruction decoding
Use table-based instruction mechanism and functions modelled like the
ARM arm specifies. This is not exactly a unified decoding framework, but
it does clean up the existing code a bit.

Converting this into a unified framework would certainly be nice to
minimize code duplication especially between alignment.c and this
emulation code, but the alignment code has been highly optimized for its
specific needs and we only need a limited set of functionality.

Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>