KVM: ARM: Change KVM_MAX_VCPUS to config option
Introduce CONFIG_KVM_ARM_MAX_VCPUS which defines the maximum number of
supported virtual CPUs per VMs on KVM/ARM. We don't want this number to
be arbitrarily large as it will take up a large amount of allocated VCPU
space for the VGIC structures for each VM. On the other hand we want to
run decent SMP VMs, and even 4 cores was until recently quite a lot on
an ARM machine, but this number is growing and for stress testing
purposes we want kernels that support more than 4 cpus, thus base the
limit on a config option.

Dynamically allocating the VGIC structures based on the number of VCPUS
will make the code even more difficult to read and understand.

Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>