HID: magicmouse: Convert to input_configured callback
In newer kernel we now have the input_configured callback, which is
executed at the same time as used to be input_register. Let's switch to
upstream's version.

This mostly repeats the following patch, except that I'm not adding
anything, but accommodating the code for newer kernels:

	commit 420359667ce49da1f5d6b081abf5cb6db0d7b700
	Author: Jaikumar Ganesh <jaikumarg@android.com>
	Date:   Tue Sep 20 16:47:43 2011 -0700

	    HID: magicmouse: Add input_register callback.

	    The magicmouse driver needs to setup the input mapping
	    after reports are parsed but before device is registered.

	    Change-Id: Ic3a0ff4fc056f5d374bb7ea1bd831a175c690b0b
	    Signed-off-by: Jaikumar Ganesh <jaikumarg@android.com>

I don't know the background for the patch, though. The initial description
isn't very helpful to understand the root cause for the change.

Signed-off-by: Anton Vorontsov <anton.vorontsov@linaro.org>