mfd: omap-usb-host: Support an auxiliary clock per port
Boards like Panda require an auxiliary clock to clock the PHY
that is connected to one of the USB ports. This patch enables
board support code to  provide the name and the rate of such
a clock for each of the USB ports. omap-usb-host driver can
then manage the clock.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>

Conflicts:

	arch/arm/plat-omap/include/plat/usb.h

Conflicts:

	arch/arm/mach-omap2/usb-host.c
	arch/arm/mach-omap2/usb.h
	drivers/mfd/omap-usb-host.c