KVM: ARM: Clenaup decode hxfar and use decode accessors
Since the decode functionality is meant to be more generic looking that
pure KVM related, we change the hxfar field to be the fault_addr field.

Also create accessors for the registers and the cpsr like we do in the
main kvm code, which should make things easier if anyone's brave enough
to implement this for arm64.

Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>