ARM: KVM: vgic: sanitize use of bitmaps
Various bits of the vgic code blindly assume that the size of u32
and unsigned long are the same. This leads to amusing side effects
on 64bit platforms, and makes the code hard to follow.

Move to proper bitmaps instead of the home grown version, and split
the vcpu specific "pending" bitmap into pending_percpu and pending_shared.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>