KVM: ARM: Convert stage2_clear_pte to unmap_stage2_range
Factor out parts of the functionality to make the code more readable and
rename to unmap_stage2_range while supporting unmapping ranges in one
go.  This paves the road for reusing this function to free the stage 2
page tables when destroying a VM; avoiding code duplication of all the
page refcounting on the page table pages themselves.

Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>