kvm tools: powerpc: convert spapr pci to new device iterators
Commit 8d35d32d0148 ("kvm tools: add generic device registration
mechanism") introduced a tree-based device lookup-by-bus mechanism as
well as iterators to enumerate the devices on a particular bus.

Whilst both x86 and ppc were converted by the original patch, the spapr
pci changes were incomplete, so include the required changes here.

Compile-tested only on ppc64 970mp. Note that I had to hack the Makefile
in order to build guest_init.o with a toolchain defaulting to ppc:

 $(GUEST_INIT): guest/init.c
        $(E) "  LINK    " $@
-       $(Q) $(CC) -static guest/init.c -o $@
-       $(Q) $(LD) -r -b binary -o guest/guest_init.o $(GUEST_INIT)
+       $(Q) $(CC) -m64 -static guest/init.c -o $@
+       $(Q) $(LD) -m elf64ppc -r -b binary -o guest/guest_init.o $(GUEST_INIT)

 $(DEPS):

Acked-by: Matt Evans <matt@ozlabs.org>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>