KVM: ARM: Clarify user_mem_abort
Wrap the decoding of the HSR to determine if this was a write fault in a
static inline so the write_fault variable can just be assigned in the
(already complicated) user_mem_abort function.

Also add comments on the icache invalidation details and on the
smp_rmb() for mmu_notifier_seq.

Cc: Will Deacon <will.deacon@arm.com>
Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>