KVM: ARM: Avodid compiler warning and bugfix
Some compilers complain about unaligned opcodes with the strings
embedded in the text section.  Therefore move the panic strings to
the rodata section to avoid the compile error.

Also, make the bad_exception point to string instead of passing the
first four characters of the string as a pointer to the panic function.

Finally, there is no need to map the panic strings into Hyp mode, as the
panic function is called from SVC mode in the host SVC address space.

Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>