KVM: ARM: vgic: Fix yet another cfg register bug
This code is so unnecessarily complicated, but it does serve the purpose
of being able to and all the bitmaps together in the end.  Oh well, we
fix another issue.

The registers are accessed by the guests view of the real HW, not the
compressed backing view, so the offsets are one-off.

Reported-by: Zhaobo (Bob, ERC) <zhaobo@huawei.com>
Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>